refactor(remote-ui): centralize remote-source state in a provider
Deferred #502 finding: every mounted useRemoteSource caller (the sidebar
section, the create-playlist modal) held its own status/playlist state and
its own waveflow:remote-changed listener, so one event fanned out N
refreshes. A RemoteSourceProvider now owns the state and the single
listener; useRemoteSource reads it from context (with an inert fallback
outside the provider). The context object + hook live in hooks/
useRemoteSource with the provider in contexts/, mirroring the
PlayerContext/usePlayer split so each file stays fast-refresh clean.
