+++ src/hooks/useRemoteSource.ts
@@ -1,9 +1,5 @@
-import { useCallback, useEffect, useRef, useState } from "react";
-import {
-  remoteGetStatus,
-  remoteListPlaylists,
-  type RemotePlaylistSummary,
-} from "../lib/tauri/remoteServer";
+import { createContext, useContext } from "react";
+import type { RemotePlaylistSummary } from "../lib/tauri/remoteServer";
 
 export interface RemoteSourceState {
   /**
@@ -11,7 +7,7 @@ export interface RemoteSourceState {
    * is bound and signed in. A stock build (feature off) makes
    * `remote_get_status` an unregistered command, which rejects and lands
    * us on `false` — the whole sidebar section then renders nothing, the
-   * same self-hiding contract as {@link RemoteServerCard}.
+   * same self-hiding contract as `RemoteServerCard`.
    */
   available: boolean;
   /** Host of the bound server, shown as the sidebar section header. */
@@ -21,84 +17,34 @@ export interface RemoteSourceState {
 }
 
 /**
- * Drives the "Remote source" sidebar section and keeps it in step with
- * the rest of the UI. Any mutation of remote data (create / rename /
- * delete a remote playlist, sign in / out) should dispatch a
- * `waveflow:remote-changed` window event; every consumer refreshes on it
- * so the sidebar, the settings card and the open view never drift apart.
+ * Shared remote-source state, owned by `RemoteSourceProvider`. Lives here
+ * (with the hook) rather than in the provider file so the provider module
+ * can stay a components-only file for fast refresh — the same split as
+ * `PlayerContext` / `usePlayer`.
+ */
+export const RemoteSourceContext = createContext<RemoteSourceState | null>(
+  null,
+);
+
+/**
+ * Read the shared remote-source state — one `waveflow:remote-changed` event
+ * refreshes every consumer once, since the provider owns the single
+ * listener. Falls back to an inert snapshot outside the provider so a stray
+ * consumer degrades to "nothing here" rather than throwing; the section is
+ * optional by design.
  */
 export function useRemoteSource(): RemoteSourceState {
-  const [available, setAvailable] = useState(false);
-  const [serverName, setServerName] = useState<string | null>(null);
-  const [playlists, setPlaylists] = useState<RemotePlaylistSummary[]>([]);
-  // Monotonic token so a slow pass can't overwrite a newer one: bursts of
-  // `waveflow:remote-changed` fire overlapping refreshes, and without this
-  // an older status/list response landing last would clobber fresher data.
-  const seqRef = useRef(0);
-  // Once we've resolved as available, sync_v2 is compiled in — so a later
-  // status failure is transient (never "command absent"), and must not
-  // make the section vanish under the user.
-  const everAvailableRef = useRef(false);
-
-  const refresh = useCallback(() => {
-    const seq = ++seqRef.current;
-    const isCurrent = () => seq === seqRef.current;
-    void (async () => {
-      try {
-        const status = await remoteGetStatus();
-        if (!isCurrent()) return;
-        if (!status.signed_in) {
-          setAvailable(false);
-          setServerName(null);
-          setPlaylists([]);
-          return;
-        }
-        everAvailableRef.current = true;
-        setAvailable(true);
-        setServerName(hostOf(status.server_url) ?? status.username ?? "Remote");
-        try {
-          const list = await remoteListPlaylists();
-          if (isCurrent()) setPlaylists(list);
-        } catch {
-          // Signed in but the list call failed (offline, transient): keep
-          // the section visible with whatever we last had rather than
-          // making it vanish under the user.
-        }
-      } catch {
-        if (!isCurrent()) return;
-        // Only hide when we've never resolved: the command may be absent
-        // (sync_v2 off). Once we've been available, this is a transient
-        // status failure — keep the last state instead of flipping off.
-        if (!everAvailableRef.current) {
-          setAvailable(false);
-          setServerName(null);
-          setPlaylists([]);
-        }
-      }
-    })();
-  }, []);
-
-  useEffect(() => {
-    refresh();
-    const onChange = () => refresh();
-    window.addEventListener("waveflow:remote-changed", onChange);
-    return () =>
-      window.removeEventListener("waveflow:remote-changed", onChange);
-  }, [refresh]);
-
-  return { available, serverName, playlists, refresh };
+  return useContext(RemoteSourceContext) ?? INERT;
 }
+
+const INERT: RemoteSourceState = {
+  available: false,
+  serverName: null,
+  playlists: [],
+  refresh: () => {},
+};
 
 /** Signal every remote-source consumer to re-read. */
 export function notifyRemoteChanged() {
   window.dispatchEvent(new Event("waveflow:remote-changed"));
-}
-
-function hostOf(url: string | null): string | null {
-  if (!url) return null;
-  try {
-    return new URL(url).host;
-  } catch {
-    return url;
-  }
 }
